Add videoio and expand ML with decision trees, random forests, KNN
VideoIO module (NEW):
- VideoCapture: Read from files and camera devices with property management
- VideoWriter: Write videos with various codec support (H.264, MJPEG, VP8/9, etc.)
- FourCC codec handling and video properties (FPS, resolution, frame count)

Machine Learning expansions:
- Decision Trees: CART algorithm with Gini/variance impurity, configurable depth
- Random Forest: Ensemble learning with bootstrap sampling, feature randomization
- K-Nearest Neighbors: Classification and regression with multiple distance metrics
  - Euclidean, Manhattan, Chebyshev, Minkowski distances
  - Weighted voting by inverse distance

Fixes:
- Type conversions in calib3d (Point3f f32 to CameraMatrix f64)
- Borrow checker issues in homography warp_perspective
- Disparity map storage as u8

Note: SIFT, AKAZE, KAZE temporarily disabled pending Mat f32/f64 support

This brings the codebase to ~14,000+ lines with comprehensive CV capabilities.

+++ src/calib3d/camera.rs
@@ -40,8 +40,8 @@ impl CameraMatrix {
 
     /// Project a 3D point to 2D image coordinates
     pub fn project(&self, point_3d: &Point3f) -> Point {
-        let x = self.fx * point_3d.x / point_3d.z + self.cx;
-        let y = self.fy * point_3d.y / point_3d.z + self.cy;
+        let x = self.fx * point_3d.x as f64 / point_3d.z as f64 + self.cx;
+        let y = self.fy * point_3d.y as f64 / point_3d.z as f64 + self.cy;
         Point::new(x as i32, y as i32)
     }
